+++ src/agenticlens/recommenders/__init__.py
@@ -1,16 +1,22 @@
 from agenticlens.recommenders.base import BaseRecommender
+from agenticlens.recommenders.chaos_impact import ChaosImpactRecommender
 from agenticlens.recommenders.duplicate_tool_calls import DuplicateToolCallsRecommender
 from agenticlens.recommenders.engine import RecommendationEngine
 from agenticlens.recommenders.excessive_chunks import ExcessiveChunksRecommender
+from agenticlens.recommenders.handoff_bloat import HandoffBloatRecommender
 from agenticlens.recommenders.long_history import LongHistoryRecommender
+from agenticlens.recommenders.rag_chunk_utility import RAGChunkUtilityRecommender
 from agenticlens.recommenders.model_swap import ModelSwapRecommender
 from agenticlens.recommenders.repeated_prompt import RepeatedSystemPromptRecommender
 
 __all__ = [
     "BaseRecommender",
+    "ChaosImpactRecommender",
     "DuplicateToolCallsRecommender",
     "ExcessiveChunksRecommender",
+    "HandoffBloatRecommender",
     "LongHistoryRecommender",
+    "RAGChunkUtilityRecommender",
     "ModelSwapRecommender",
     "RecommendationEngine",
     "RepeatedSystemPromptRecommender",